feat: controladores y vistas

+++ Acme-Newspaper/src/main/java/services/ArticleService.java
@@ -2,6 +2,7 @@ package services;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Date;
 
 import javax.transaction.Transactional;
 
@@ -11,6 +12,7 @@ import org.springframework.util.Assert;
 
 import repositories.ArticleRepository;
 import domain.Article;
+import domain.User;
 
 @Service
 @Transactional
@@ -25,6 +27,9 @@ public class ArticleService {
 
 	@Autowired
 	private UserService userService;
+	
+	@Autowired
+	private AdministratorService administratorService;
 
 	// Constructor ------------------------------------------------------------
 
@@ -37,7 +42,16 @@ public class ArticleService {
 	public Article create() {
 		this.userService.checkAuthority();
 		Article result;
+		Date moment;
+		User writer;
+		
 		result = new Article();
+		writer = userService.findByPrincipal();
+		
+		moment = new Date(System.currentTimeMillis() - 1000);
+		result.setMoment(moment);
+		result.setWriter(writer);
+		
 		return result;
 	}
 
@@ -56,8 +70,9 @@ public class ArticleService {
 
 	public Article save(final Article article) {
 		this.userService.checkAuthority();
-		Assert.isTrue(this.userService.findArticleCreator(article.getWriter()
-				.getId()) == this.userService.findByPrincipal());
+		if (article.getId() != 0) {
+			Assert.isTrue(article.getDraftmode() == false);
+		}
 		Article result = article;
 		Assert.notNull(article);
 		result = this.articleRepository.save(result);
@@ -65,6 +80,7 @@ public class ArticleService {
 	}
 
 	public void delete(final Article article) {
+		this.administratorService.checkAuthority();
 		Assert.notNull(article);
 		Assert.isTrue(article.getId() != 0);
 		this.articleRepository.delete(article);
@@ -76,12 +92,6 @@ public class ArticleService {
 		Collection<Article> res;
 		res = this.articleRepository.findArticleByNewspaper(id);
 		return res;
-	}
-	
-	public Article search() {
-		Article result;
-		result = new Article();
-		return result;
 	}
 	
 	public Collection<Article> searchArticle(String criteria) {
