+++ Acme-Newspaper/src/main/java/services/NewspaperService.java
@@ -2,14 +2,18 @@ package services;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
+import org.springframework.validation.BindingResult;
 
 import repositories.NewspaperRepository;
+import domain.Article;
 import domain.Newspaper;
+import domain.User;
 
 @Service
 @Transactional
@@ -21,6 +25,15 @@ public class NewspaperService {
 	private NewspaperRepository newspaperRepository;
 
 	// Supporting services ----------------------------------------------------
+	
+	@Autowired
+	private UserService userService;
+	
+	@Autowired
+	private AdministratorService administratorService;
+	
+	@Autowired
+	private ArticleService articleService;
 
 	// Constructor ------------------------------------------------------------
 
@@ -31,8 +44,21 @@ public class NewspaperService {
 	// Simple CRUD methods ----------------------------------------------------
 
 	public Newspaper create() {
+		this.userService.checkAuthority();
 		Newspaper result;
+		User publisher;
+		Date publication;
+		Collection<Article> articles;
+		
+		publisher = userService.findByPrincipal();
 		result = new Newspaper();
+		articles = new ArrayList<Article>();
+		
+		publication = new Date(System.currentTimeMillis() - 1000);
+		result.setPublication(publication);
+		result.setPublisher(publisher);
+		result.setArticles(articles);
+		
 		return result;
 	}
 
@@ -52,6 +78,7 @@ public class NewspaperService {
 	}
 
 	public Newspaper save(Newspaper newspaper) {
+		this.userService.checkAuthority();
 		Assert.notNull(newspaper);
 		Newspaper res;
 		res = this.newspaperRepository.save(newspaper);
@@ -59,9 +86,14 @@ public class NewspaperService {
 	}
 
 	public void delete(Newspaper newspaper) {
+		this.administratorService.checkAuthority();
 		Assert.notNull(newspaper);
 		Assert.isTrue(newspaper.getId() != 0);
 		Assert.isTrue(this.newspaperRepository.exists(newspaper.getId()));
+		Collection<Article> articles;
+		articles = new ArrayList<Article>(this.articleService.findArticleByNewspaper(newspaper.getId()));
+		for (final Article s : articles)
+			this.articleService.delete(s);
 		this.newspaperRepository.delete(newspaper);
 	}
 
@@ -73,8 +105,41 @@ public class NewspaperService {
 		return res;
 	}
 	
+	public Collection<Newspaper> findNewspapersPublicated() {
+		Collection<Newspaper> res = new ArrayList<Newspaper>();
+		res.addAll(newspaperRepository.findNewspapersPublicated());
+		return res;
+	}
+	
+	public Collection<Newspaper> findNewspapersNotPublicated() {
+		Collection<Newspaper> res = new ArrayList<Newspaper>();
+		res.addAll(newspaperRepository.findNewspapersNotPublicated());
+		return res;
+	}
+	
 	public void flush() {
 		this.newspaperRepository.flush();
+	}
+	
+	public Newspaper reconstruct(final Newspaper newspaper, final BindingResult binding) {
+		Newspaper res;
+		Newspaper newspaperFinal;
+		if (newspaper.getId() == 0)
+			res = newspaper;
+		else {
+			newspaperFinal = this.findOne(newspaper.getId());
+			newspaper.setPublisher(this.userService.findByPrincipal());
+			newspaperFinal.setTitle(newspaper.getTitle());
+			newspaperFinal.setPublication(newspaper.getPublication());
+			newspaperFinal.setDescription(newspaper.getDescription());
+			newspaperFinal.setPicture(newspaper.getPicture());
+			newspaperFinal.setHide(newspaper.getHide());
+			newspaperFinal.setArticles(newspaper.getArticles());
+			newspaperFinal.setPublisher(newspaper.getPublisher());
+			res = newspaperFinal;
+		}
+
+		return res;
 	}
 
 }
